# src/main/gen-src/com/dcs/user/registration/model/RegistrationData.java
/*
 * User Registation API
 * User Registation API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: test@test.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.dcs.user.registration.model;

import com.dcs.user.registration.exceptions.EventPlanningRuntimeException;
import org.springframework.util.StringUtils;

/**
 * RegistrationData
 */
public record RegistrationData(String email, String password, String mobileNumber) {

	public RegistrationData
	{
		if (!StringUtils.hasLength(email) || !StringUtils.hasLength(password))
		{
			throw new EventPlanningRuntimeException("Filed can not be empty or null.");
		}
	}
}